Guess names using numbered name strings
Numbered names (e.g. ClimateDefines_0) are common in AAMP files,
because the format lacks support for proper arrays.

These are put in a file that is separate from the regular
hashed_names.txt to avoid having to generate hundreds of entries
that differ only in the number and to avoid making having a huge list.