Fix data corruption bug in parameter data reuse logic
Fixes a data corruption bug in the parameter data reuse logic that
could occur when a match with an offset that is not divisible by 4
is found. The offset writer would round down the offset to the
nearest multiple of 4, which results in data corruption.

This also adds offset representability checks to the placeholder
offset writers in order to catch similar bugs.

A test case has been added.

Fixes #7